refactor(react-native): split polyfills into opt-in imports, wire LicenseContext
Address PR feedback from @AlemTuzlak:

- Split monolithic polyfills.ts into individual imports (streams, encoding,
  crypto, dom, location) so users control what gets polyfilled
- Remove auto-polyfill from index.ts — users explicitly import what they need
- Add sideEffects array to package.json for bundler correctness
- Wire LicenseContext.Provider into the RN CopilotKitProvider
- Add new polyfill entry points to tsdown config and package.json exports
- Use globalThis consistently (no more declare const global conflicts)
