+++ packages/react-native/tsdown.config.ts
@@ -1,7 +1,15 @@
 import { defineConfig } from "tsdown";
 
 export default defineConfig({
-  entry: ["src/index.ts", "src/polyfills.ts"],
+  entry: [
+    "src/index.ts",
+    "src/polyfills.ts",
+    "src/polyfills/streams.ts",
+    "src/polyfills/encoding.ts",
+    "src/polyfills/crypto.ts",
+    "src/polyfills/dom.ts",
+    "src/polyfills/location.ts",
+  ],
   format: ["esm", "cjs"],
   dts: true,
   sourcemap: true,
